Pass error instead of JSON data.
This commit allows us to pass the error (returned from Github) as an actual JS object (and not a
JSON-formatted string). A reference is also attached to the error so we can see all of the JSON-data
that was originally returned from Github (when the request finished). The specifications for this
change was requested by @rvagg here: https://github.com/rvagg/ghauth/issues/12#issuecomment-68654454.